+++ database/migrations/2018_02_19_101615_create_quotes_table.php
@@ -19,7 +19,7 @@ class CreateQuotesTable extends Migration
             $table->integer('client_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('client_id')->references('id')->on('clients')->onDelete('cascade');
-            $table->integer('delivery_price');
+            $table->decimal('delivery_price', 12, 2);
             $table->text('note');
             $table->timestamps();
         });
